// src/app/Models/ProductoModel.ts
export interface ProductoModel {
    Id_Producto?: number;
    NombreProducto?: string;
    Descripcion?: string;
    Tipo?: string;
    Precio?: number;
    Cantidad?: number;
    Id_Usuario?: number;
    Autorizado?: string;
    CantidadVendidos?: number;
    Id_Autorizador?: number;
    Valoracion?: number;
    Registro?: string;
    Estado?:string;
  }
  